+++ examples/scripts/metal_dielectric_metal_approximations.py
@@ -1,4 +1,4 @@
-'''Example metal-insulator-metal dispersion approximations example usage'''
+'''Example metal-dielectric-metal dispersion approximations example usage'''
 
 import numpy as np
 import matplotlib
@@ -10,24 +10,24 @@ from optical_dispersion_relations import plasmon
 matplotlib.rc('font', size=12)
 
 collin_approximation_parameters = {
+    'wavelength': 1,
+    'thickness': np.logspace(-2, 1, 100),
     'dielectric_permittivity': 1,
     'metal_permittivity': -50,
-    'wavelength': 1,
-    'insulator_thickness': np.logspace(-2, 1, 100),
 }
 
 sondergaard_approximation_parameters = {
+    'wavelength': 775,
+    'thickness': np.arange(1, 3000),
     'dielectric_permittivity': 1,
     'metal_permittivity': -23.6+1.69j,
-    'wavelength': 775,
-    'insulator_thickness': np.arange(1, 3000),
 }
 
-collin_approximation = plasmon.metal_insulator_metal_collin_approximation(
+collin_approximation = plasmon.metal_dielectric_metal_collin_approximation(
     **collin_approximation_parameters
 )
 
-sondergaard_approximation = plasmon.metal_insulator_metal_sondergaard_narrow_approximation(
+sondergaard_approximation = plasmon.metal_dielectric_metal_sondergaard_narrow_approximation(
     **sondergaard_approximation_parameters
 )
 
@@ -39,7 +39,7 @@ sondergaard_approximation_figure, (sondergaard_approximation_axes) = plt.subplot
 collin_approximation_axes.set_title(
     'Fig. 2 from \nhttps://doi.org/10.1364/OE.15.004310')
 collin_approximation_axes.semilogx(
-    collin_approximation_parameters['insulator_thickness'],
+    collin_approximation_parameters['thickness'],
     collin_approximation.real,
     color='C3',
     label='coupled SPP approx.')
@@ -51,12 +51,12 @@ collin_approximation_axes.set_ylabel(r'Effective index $n_{1D}$')
 collin_approximation_axes.legend()
 collin_approximation_figure.tight_layout()
 collin_approximation_figure.savefig(
-    './images/collin_approximation_example.png')
+    '../images/collin_approximation_example.png')
 
 sondergaard_approximation_axes.set_title(
     'Fig. 4 from \nhttps://doi.org/10.1364/OE.15.010869')
 sondergaard_approximation_axes.plot(
-    sondergaard_approximation_parameters['insulator_thickness'],
+    sondergaard_approximation_parameters['thickness'],
     sondergaard_approximation.real,
     color='C2',
     label='small-gap approximation')