+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Exact dispersion relations for MIM and IMI plasmonic systems:
+	* `transcendential_trilayer_even_magnetic_field`
+
 ### Changed
 
 * BREAKING: Change DrudeLorentz default plasma frequency to 1.