+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Exact dispersion relations for MIM and IMI plasmonic systems:
+* Transcendential equations for trilayer plasmonic dispersion relations:
 	* `transcendential_trilayer_even_magnetic_field`
 	* `transcendential_trilayer_odd_magnetic_field`